Add Market Data and Create a file that predicts the stock price difference between opening and closing on any given day
add the excel file that includes the data that is to be used for the model.

create a python file with the classes get_data (which imports, splits and prepares the data for training), visualize_data (creates graphs to help better visualize the data), train_model (trains the model), and test_model (tests the model).